Add tests of OSSkipCondition on Mac and Linux
